github-repo-chatbot/requirements.txt
fastapi
uvicorn
gitpython
langchain
langchain-community
langchain-openai
langchain-qdrant
sentence-transformers
groq
qdrant-client
python-dotenv